## Proposed changes

We can avoid double encoding of messages published to Nats by using the client protocol v2 push format. Thus saving one alloc. The format of the messages we publish is an internal implementation detail, so we can change it.

Before:

```
BenchmarkNatsEnginePublish-12              	 3881850	       296.7 ns/op	      56 B/op	       2 allocs/op
BenchmarkNatsEnginePublishParallel-12      	 3006904	       416.1 ns/op	      56 B/op	       2 allocs/op
```

After (-1 alloc)

```
BenchmarkNatsEnginePublish-12              	 4523809	       254.9 ns/op	      32 B/op	       1 allocs/op
BenchmarkNatsEnginePublishParallel-12      	 3242451	       363.9 ns/op	      32 B/op	       1 allocs/op
``` 

Also, will be slightly faster on message unmarshaling when receiving it from PUB/SUB.